Move drawable object piece into default.
